ShopManagement/src/main/java/com/ivan/ShopManagement/model/Statistics.java
package com.ivan.ShopManagement.model;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;

@Entity
@Table(name = "statistics")
public class Statistics {

    @Id
    @Column(name = "id")
    private long id;

    @Column(name = "product")
    private String product;

    @Column(name = "quantity")
    private int quantity;

    public Statistics(){

    }

    public Statistics(String product, int quantity) {
        this.product = product;
        this.quantity = quantity;
    }

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }

    public String getProduct() {
        return product;
    }

    public void setProduct(String product) {
        this.product = product;
    }

    public int getQuantity() {
        return quantity;
    }

    public void setQuantity(int quantity) {
        this.quantity = quantity;
    }
}
